Blood-Bank-Management-System/requirements.txt
Python==3.12 
Flask==2.2.5
mysql-connector-python==8.2.0